+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - 券種検索条件をオファー検索条件に統合
+- ticketTypesコレクションをoffersコレクションへ移行
 
 ### Deprecated